Updates timber 2 code with thinktimber additions
