--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,35 +1,0 @@
-<!--
-First off, hello!
-
-Thanks for submitting a PR. We love/welcome PRs (especially if it's your first).
--->
-
-<!-- Remove this if no related tickets exist. -->
-<!-- You can add the related ticket numbers here using #. Example: #2471 -->
-Related:
-
-- Ticket 1
-- Ticket 2
-
-## Issue
-<!-- Description of the problem that this code change is solving -->
-
-
-## Solution
-<!-- Description of the solution that this code changes are introducing to the application. -->
-
-
-## Impact
-<!-- What impact will this have on the current codebase, performance, backwards compatibility? -->
-
-
-## Usage Changes
-<!-- Are there are any usage changes that we need to know about? If so, list them here so that we can integrate it in the release notes and developers know what usage changes are associated to your PR.
--->
-
-## Considerations
-<!-- As we do not live in an ideal world it's worth to share your thought on how we could make the solution even better. -->
-
-
-## Testing
-<!-- Are unit tests included? If they need to be written, please provide pseudo code for a scenario that fails without your code, but succeeds with it -->
