+++ 404.php
@@ -8,4 +8,4 @@ namespace App;
 use Timber\Timber;
 
 $context = Timber::context();
-Timber::render( 'templates/404.twig', $context );
+Timber::render( 'pages/404.twig', $context );
