+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "thinkwp",
+  "name": "thinktimber",
   "version": "1.0.0",
-  "description": "thinkwp",
+  "description": "A ThinkShout WordPress theme.",
   "devDependencies": {
     "@babel/core": "^7.4.4",
     "@babel/preset-env": "^7.4.4",
