--- views/partials/comment.twig
@@ -1,19 +1,0 @@
-<div class="blog-comment" id="blog-comment-{{ comment.id }}">
-    <h5 class="comment-author">{{ comment.author.name }} says</h5>
-    <div class="comment-content">{{ comment.content|wpautop }}</div>
-
-    <section class="comment-box">
-        {% include 'partials/comment-form.twig' %}
-
-        {% if post.comments %}
-            <h4>replies</h4>
-            <div class="comments">
-                {% for cmt in comment.children %}
-                    {% include 'partials/comment.twig' with {
-                        comment: cmt
-                    } %}
-                {% endfor %}
-            </div>
-        {% endif %}
-    </section>
-</div>
