--- views/partials/footer.twig
@@ -1,1 +1,0 @@
-<footer id="footer">Copyright {{ 'now'|date('Y') }}</footer>
