--- views/partials/head.twig
@@ -1,9 +1,0 @@
-<head>
-    <meta charset="{{ site.charset }}" />
-    <link rel="stylesheet" href="{{ site.theme.link }}/style.css" type="text/css" media="screen" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="author" href="{{ site.theme.link }}/humans.txt" />
-    <link rel="profile" href="http://gmpg.org/xfn/11" />
-    {% do action('get_header') %}
-    {{ function('wp_head') }}
-</head>
