--- views/partials/menu.twig
@@ -1,12 +1,0 @@
-{% if menu %}
-    <ul>
-        {% for item in items %}
-            <li class="{{ item.classes|join(' ') }}">
-                <a target="{{ item.target }}" href="{{ item.link }}">{{ item.title }}</a>
-                {% include 'partials/menu.twig' with {
-                    items: item.children
-                } %}
-            </li>
-        {% endfor %}
-    </ul>
-{% endif %}
