--- views/partials/tease-post.twig
@@ -1,16 +1,0 @@
-{% extends 'partials/tease.twig' %}
-
-{% block content %}
-    <h2 class="h2"><a href="{{ post.link }}">{{ post.title }}</a></h2>
-    <p>
-        {{
-            post.excerpt({
-                words: 5,
-                read_more: 'Keep reading'
-            })
-        }}
-    </p>
-    {% if post.thumbnail.src %}
-        <img src="{{ post.thumbnail.src }}" />
-    {% endif %}
-{% endblock %}
