--- views/partials/tease.twig
@@ -1,9 +1,0 @@
-<article class="tease tease-{{ post.type }}" id="tease-{{ post.id }}">
-    {% block content %}
-        <h2 class="h2"><a href="{{ post.link }}">{{ post.title }}</a></h2>
-        <p>{{ post.excerpt }}</p>
-        {% if post.thumbnail %}
-            <img src="{{ post.thumbnail.src }}" />
-        {% endif %}
-    {% endblock %}
-</article>
