--- views/templates/404.twig
@@ -1,5 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    Sorry, we couldn't find what you're looking for.
-{% endblock %}
