--- views/templates/archive.twig
@@ -1,15 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    {% for post in posts %}
-        {% include ['partials/tease-' ~ post.type ~ '.twig', 'partials/tease.twig'] %}
-    {% endfor %}
-
-    {% include 'partials/pagination.twig' with {
-        pagination: posts.pagination({
-            show_all: false,
-            mid_size: 3,
-            end_size: 2
-        })
-    } %}
-{% endblock %}
