--- views/templates/author.twig
@@ -1,7 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    {% for post in posts %}
-        {% include ['partials/tease-' ~ post.type ~ '.twig', 'partials/tease.twig'] %}
-    {% endfor %}
-{% endblock %}
