--- views/templates/index.twig
@@ -1,16 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    <h2>{{ foo }}</h2>
-    {% for post in posts %}
-        {% include ['partials/tease-' ~ post.type ~ '.twig', 'partials/tease.twig'] %}
-    {% endfor %}
-
-    {% include 'partials/pagination.twig' with {
-        pagination: posts.pagination({
-            show_all: false,
-            mid_size: 3,
-            end_size: 2
-        })
-    } %}
-{% endblock %}
