--- views/templates/page.twig
@@ -1,12 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    <div class="content-wrapper">
-        <article class="post-type-{{ post.type }}" id="post-{{ post.id }}">
-            <section class="article-content">
-                <h1 class="article-h1">{{ post.title }}</h1>
-                <div class="article-body">{{ post.content }}</div>
-            </section>
-        </article>
-    </div>
-{% endblock %}
