--- views/templates/search.twig
@@ -1,18 +1,0 @@
-{# see `templates/archive.twig` for an alternative strategy of extending templates #}
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    <div class="content-wrapper">
-        {% for post in posts %}
-            {% include ['partials/tease-' ~ post.type ~ '.twig', 'partials/tease.twig'] %}
-        {% endfor %}
-
-        {% include 'partials/pagination.twig' with {
-            pagination: posts.pagination({
-                show_all: false,
-                mid_size: 3,
-                end_size: 2
-            })
-        } %}
-    </div>
-{% endblock %}
