--- views/templates/single-password.twig
@@ -1,17 +1,0 @@
-{% extends 'layouts/base.twig' %}
-
-{% block content %}
-    <form class="password-form"
-        action="{{ site.link }}/wp-login.php?action=postpass"
-        method="post">
-        <label for="pwbox-{{ post.id }}">Password:</label>
-        <input class="password-box"
-            name="post_password"
-            id="pwbox-{{ post.id }}"
-            type="password"
-            placeholder="Password"
-            size="20"
-            maxlength="20" />
-        <input class="password-btn" type="submit" name="Submit" value="Submit" />
-    </form>
-{% endblock %}
